dhcp: Refactor DHCP bits out
